back_pizzeria/requirements.txt
asgiref==3.2.10
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.4.1
